Time chart action reducer bug fix
